Update plotVisRespSAT.m
Now computing both latency and magnitude of the visual response. Taking magnitude currently at the average value of the VR for 100 ms post-visual response onset. This means that the intervals over which this average is taken will be difference in the Fast and Accurate conditions, and across neurons. There are some units for which magAcc > magFast.