Developed code for analysis of visual response
Added new struct NSTATS. This struct contains summary statistics of the spiking activity at different epochs (baseline, visual response, post-saccade, etc.). The idea is to maintain consistent stats across all analyses for the paper. When a function computes a stat, such as visual response magnitude, that stat is added to the NSTATS struct, and is there to be accessed for future analyses and for reference. Also updated analysis of visual response, specifically trial-to-trial change in VR. Note that the prior analysis of VR was insufficient for a couple reasons: (1) There was no way of testing for a significant difference in VR magnitude across conditions, and (2) The normalization factor for VR magnitude was arbitrary, and different from the norm factor used on the SDF. The next step is to confirm use of an appropriate norm factor, and complete test for significant difference in VR magnitude.