Updated fxns for VR magnitude analysis
Updated testVRMagSATeffect() to test for the existence of an effect of SAT condition on the VR on a per-cell basis. This is based on spike counts in an interval 100-ms post-onset of the VR. The combination of this with calculation of the SDF-based magnitude needs to be resolved, as some cells that are tagged as Acc>Fast actually show greater SDF in Fast, and vice versa. This will complete Fig. 3 (visual response).