Updated analyses of behavior (Fig. 1), baseline activity (Fig. 2), and visual response (Fig. 3)
Completed analyses for Supp. Fig. 1, including ISI individually for each monkey. Currently in the process of computing TST for those visually-responsive neurons that select target amongst distractors. For this we need SDF's for Target In and Distractor In. It appears that roughly half of the SEF visual population does target selection. It also appears that TST_Fast < TST_Acc, but need to run the stats.