Updated analysis of Vis Response and Choice Error activity
Updated ability to detect the choice error-related (post-primary response) signal. This is based on two measures: (1) significant difference between Correct and Error SDFs, and (2) minimum magnitude difference based on the baseline period. Also have code in place to determine neurons for which time of onset of error encoding is explained by ISI of the secondary saccade. These neurons will be excluded as not error-related. Currently using RT-matching across correct and error trials.